Feat: Member Validate, Authentication

+++ src/main/java/com/example/demo/web/MemberController.java
@@ -12,10 +12,15 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import java.util.Objects;
 
 @Controller
 @RequestMapping("/members")
@@ -49,28 +54,45 @@ public class MemberController {
     }
 
     @PostMapping("/add")
-    public String add(@ModelAttribute MemberAddDto memberAddDto) {
+    public String add(@Validated @ModelAttribute MemberAddDto memberAddDto, BindingResult bindingResult) {
+        if (bindingResult.hasErrors()) {
+            log.info("errors={}", bindingResult);
+            return "/members/add";
+        }
         Member saveMember = memberService.save(memberAddDto);
         log.info("saveMember(Controller)= {} ", saveMember);
         return "redirect:/";
     }
 
     @GetMapping("/{memberId}/edit")
-    public String editMember(@PathVariable Long memberId, Model model) {
+    public String editMember(@PathVariable Long memberId, Authentication authentication,RedirectAttributes redirectAttributes,Model model) {
         MemberGetDto member = memberService.getMember(memberId);
+        if (!Objects.equals(member.getLoginId(), authentication.getName())) {
+            redirectAttributes.addFlashAttribute("flag", true);
+            return "redirect:/members/{memberId}";
+        }
         model.addAttribute("member", member);
         return "/members/edit";
     }
 
     @PostMapping("/{memberId}/edit")
-    public String edit(@PathVariable Long memberId, @ModelAttribute("member") MemberUpdateDto memberUpdateDto, RedirectAttributes redirectAttributes) {
+    public String edit(@PathVariable Long memberId,@Validated @ModelAttribute("member") MemberUpdateDto memberUpdateDto,BindingResult bindingResult, RedirectAttributes redirectAttributes) {
+        if (bindingResult.hasErrors()) {
+            log.info("errors={}", bindingResult);
+            return "/members/edit";
+        }
         memberService.updateMember(memberId, memberUpdateDto);
         redirectAttributes.addAttribute("memberId", memberId);
         return "redirect:/members/{memberId}";
     }
 
     @PostMapping("/{memberId}/delete")
-    public String delete(@PathVariable Long memberId) {
+    public String delete(@PathVariable Long memberId, Authentication authentication,RedirectAttributes redirectAttributes) {
+        MemberGetDto member = memberService.getMember(memberId);
+        if (!Objects.equals(member.getLoginId(), authentication.getName())) {
+            redirectAttributes.addFlashAttribute("flag", true);
+            return "redirect:/members/{memberId}";
+        }
         memberService.deleteMember(memberId);
         return "redirect:/members";
     }
